Actually return client response channel
Current behaviour is to return a closed channel from client.TxResponses.
Receives from the returned channel always block. The impact is that we
never print debug responses when beeline is configured in debug mode.

I assume that this was not the original intention of the function. I
assume that the original intention of the function is to actually return
the underlying client responses channel, if the client is not nil. This
change matches that assumed intention.

It allows us to actually get debug responses when beeline is configured
to run in debug mode.